+++ main.py
@@ -1,19 +1,8 @@
 from openai import OpenAI
 from rich.console import Console
-from logic import *
 from handlers import *
 from storage import *
-
-# ────────────────────────────────────────────────────────────────────────────
-# Optional OpenAI (autocorrect & semantic search); can work without key.txt
-# ────────────────────────────────────────────────────────────────────────────
-try:
-    from openai import OpenAI
-
-    with open("key.txt", "r", encoding="utf‑8") as f:
-        _client = OpenAI(api_key=f.read().strip())
-except (ImportError, FileNotFoundError):
-    _client = None
+from openai_file import _client
 
 def main():
     users = load_users()
